+++ css/main.css
@@ -1,0 +1,5 @@
+@import "base.css";
+@import "logs.css";
+@import "modals.css";
+@import "toolbar.css";
+@import "settings.css";
