+++ index.html
@@ -7,8 +7,8 @@
     <title>Robot</title>
     <link rel="stylesheet" href="node_modules/material-symbols/index.css" />
     <link rel="stylesheet" href="node_modules/bulma/css/bulma.css" />
-    <link rel="stylesheet" href="src/style.css" />
-    <link rel="shortcut icon" href="src/favicon.ico" />
+    <link rel="stylesheet" href="css/main.css" />
+    <link rel="shortcut icon" href="images/favicon.ico" />
   </head>
 
   <body>
@@ -227,7 +227,7 @@
             <div class="settings-tab is-hidden" id="settings-about">
               <div class="columns">
                 <div class="column has-text-centered">
-                  <img src="src/logo.svg" alt="Student Robotics logo" width="65%">
+                  <img src="images/logo.svg" alt="Student Robotics logo" width="65%">
                 </div>
                 <div class="column is-align-content-center">
                   <strong class="is-size-5 mb-2" data-from-metadata="os_pretty_name"></strong>
@@ -266,6 +266,6 @@
       </div>
     </template>
 
-    <script src="src/main.js" type="module"></script>
+    <script src="js/main.js" type="module"></script>
   </body>
 </html>
