+++ js/settings.js
