+++ package.json
@@ -7,8 +7,8 @@
     "start:lan": "vite --host 0.0.0.0",
     "build": "vite build",
     "preview": "vite preview",
-    "format": "prettier --write src/",
-    "lint": "prettier --check src/"
+    "format": "prettier --write images/",
+    "lint": "prettier --check images/"
   },
   "repository": {
     "type": "git",
