--- src/style.css
@@ -1,5 +1,0 @@
-@import "css/base.css";
-@import "css/logs.css";
-@import "css/modals.css";
-@import "css/toolbar.css";
-@import "css/settings.css";
